Moved also MCO to separate plugin
