+++ force_bdss/core_plugins/test_mco/basic/basic_optimizer.py
@@ -1,13 +1,13 @@
 import subprocess
-
 import sys
 
-from force_bdss.mco.base_multi_criteria_optimizer import \
+from force_bdss.mco.base_multi_criteria_optimizer import (
     BaseMultiCriteriaOptimizer
+)
 
 
-class DakotaOptimizer(BaseMultiCriteriaOptimizer):
+class BasicOptimizer(BaseMultiCriteriaOptimizer):
     def run(self):
-        print("Running dakota optimizer")
+        print("Running Basic optimizer")
         subprocess.check_call([sys.argv[0], "--evaluate",
                                self.application.workflow_filepath])
