+++ force_bdss/core_plugins/test_mco/dakota/dakota_optimizer.py
@@ -1,12 +1,14 @@
 import subprocess
+
 import sys
 
-from force_bdss.mco.base_multi_criteria_optimizer import \
+from force_bdss.mco.base_multi_criteria_optimizer import (
     BaseMultiCriteriaOptimizer
+)
 
 
-class BasicOptimizer(BaseMultiCriteriaOptimizer):
+class DakotaOptimizer(BaseMultiCriteriaOptimizer):
     def run(self):
-        print("Running Basic optimizer")
+        print("Running dakota optimizer")
         subprocess.check_call([sys.argv[0], "--evaluate",
                                self.application.workflow_filepath])
